requirements.txt

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0
pillow>=9.5.0
scikit-learn>=1.5.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
ftfy>=6.1.0
regex>=2023.0.0
IQA-pytorch>=0.1.0
einops>=0.6.0
tqdm>=4.65.0
omegaconf>=2.3.0
wandb>=0.15.0 
opencv-python>=4.8.0
scikit-image>=0.21.0
matplotlib>=3.7.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
